refactor(prefer-item): 기관 선호 물품 조회 리팩토링
