+++ src/main/java/com/somemore/center/usecase/PreferItemQueryUseCase.java
@@ -1,12 +1,12 @@
-package com.somemore.domains.center.usecase.query;
+package com.somemore.center.usecase;
 
-import com.somemore.domains.center.domain.PreferItem;
-import com.somemore.domains.center.dto.response.PreferItemResponseDto;
-
+import com.somemore.center.domain.PreferItem;
+import com.somemore.center.dto.response.PreferItemResponseDto;
 import java.util.List;
 import java.util.UUID;
 
 public interface PreferItemQueryUseCase {
+
     List<PreferItemResponseDto> getPreferItemDtosByCenterId(UUID centerId);
 
     List<PreferItem> getPreferItemsByCenterId(UUID centerId);
